backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "fastapi[standard,starndard]>=0.135.0",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "pyjwt>=2.10.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.35.0",
]
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "pytest<8.0.0,>=7.4.3",
    "mypy",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
